separated ble, encryption, wifi, and soil assessment functions into their own library files. Created a global pin configure file and updated the utilities file to contain aes keys
