+++ .gitignore
@@ -5,3 +5,4 @@
 .vscode/ipch
 
 utilities.h
+utilities.cpp
