+++ lib/BleHandler/BleHandler.h
@@ -1,0 +1,7 @@
+#ifndef BLE_HANDLER_H
+#define BLE_HANDLER_H
+
+void turnOffBle();
+void beginBLE();
+
+#endif
