+++ lib/ButtonHandler/ButtonHandler.h
@@ -1,0 +1,10 @@
+#ifndef ButtonHandler_h
+#define ButtonHandler_h
+
+#include <Arduino.h>
+
+void setupStatusButtonInterrupt();
+void attachStatusButtonInterrupt();
+void IRAM_ATTR statusBtnTriggered();
+
+#endif
