+++ lib/Encryption/Encryption.h
@@ -1,0 +1,7 @@
+#ifndef ENCRYPTION_H
+#define ENCRYPTION_H
+
+void encryptAndStore(const char* ssid, const char* password);
+void loadAndDecrypt();
+
+#endif
