+++ lib/LedControl/LedControlConstants.h
@@ -1,7 +1,6 @@
 #ifndef LED_CONTROL_CONSTANTS_H
 #define LED_CONTROL_CONSTANTS_H
 
-
 #define PWM1_Ch    0
 #define PWM1_Res   8
 #define PWM1_Freq  1000
@@ -12,5 +11,9 @@
 #define PWM3_Res   8
 #define PWM3_Freq  1000
 
+namespace ColorSettings {
+    const int GREEN[3] = {0, 255, 0};
+    const int RED[3] = {255, 0, 0};
+}
 
 #endif
