+++ lib/MqttHandler/MqttHandler.h
@@ -1,0 +1,18 @@
+#include "SoilSensor.h"
+#ifndef MQTT_HANDLER_H
+#define MQTT_HANDLER_H
+
+extern bool shadow_auto;
+extern bool shadow_pump;
+extern bool checkMQTTShadow; 
+extern bool checkInterupt;
+void getAndCheckShadowState();
+void updateShadowReportedState();
+void logSoilSensor(soilSensorResponse currentSoilResponse, bool didWater);
+void messageHandler(String &topic, String &payload);
+void connectToMQTT();
+void mqttLoop();
+bool checkMqttStatus();
+void retrieveShadowOnMqttConnection();
+
+#endif
