+++ lib/WifiHandler/WifiHandler.h
@@ -1,0 +1,9 @@
+#ifndef WIFI_HANDLER_H
+#define WIFI_HANDLER_H
+
+extern const char* ssid;      
+extern const char* password; 
+void connectToWiFi();
+bool checkWifiStatus();
+
+#endif
