+++ lib/WifiHandler/WifiHandlerConstants.h
@@ -1,0 +1,8 @@
+#ifndef WIFI_HANDLER_CONSTANTS_H
+#define WIFI_HANDLER_CONSTANTS_H
+
+namespace WifiSettings {
+    const int MAX_RETRIES = 10;
+}
+
+#endif
